Extract common IFF code to separate module.
Update the AIFF and RIFF implementations to use the new IFF base classes.
